Fixing tests.

+++ internal/app/aggregation/aggregator_test.go
@@ -55,7 +55,7 @@ func TestItemEventsAggregator(t *testing.T) {
 			}
 
 			fetchResultChan := make(chan fetchMessageResult)
-			mockModel.EXPECT().fetchMessages(ctx, 0).Return(fetchResultChan)
+			mockModel.EXPECT().fetchMessages(ctx, int64(0)).Return(fetchResultChan)
 
 			cnt := newCounters()
 
@@ -115,7 +115,7 @@ func TestItemEventsAggregator(t *testing.T) {
 			mockModel, _ := deps.deps.AggregatorModel.(*mockItemEventsAggregatorModel)
 
 			fetchResultChan := make(chan fetchMessageResult)
-			mockModel.EXPECT().fetchMessages(ctx, 0).Return(fetchResultChan)
+			mockModel.EXPECT().fetchMessages(ctx, int64(0)).Return(fetchResultChan)
 			cnt := newCounters()
 
 			exit := make(chan error)
@@ -136,7 +136,7 @@ func TestItemEventsAggregator(t *testing.T) {
 			fetchResultChan := make(chan fetchMessageResult)
 
 			mockModel, _ := deps.deps.AggregatorModel.(*mockItemEventsAggregatorModel)
-			mockModel.EXPECT().fetchMessages(ctx, 0).Return(fetchResultChan)
+			mockModel.EXPECT().fetchMessages(ctx, int64(0)).Return(fetchResultChan)
 			cnt := newCounters()
 
 			exit := make(chan error)
@@ -156,7 +156,7 @@ func TestItemEventsAggregator(t *testing.T) {
 			cnt := newCounters()
 
 			fetchResultChan := make(chan fetchMessageResult)
-			mockModel.EXPECT().fetchMessages(ctx, 0).Return(fetchResultChan)
+			mockModel.EXPECT().fetchMessages(ctx, int64(0)).Return(fetchResultChan)
 			mockModel.EXPECT().flushMessages(ctx, cnt)
 
 			exit := make(chan error)
